LocUtil occasionally failed due to XmlInitialization
Split into separate classes to prevent potential parallelism race conditions.
